Enable the use of the OverlayFS for the LiveOS root filesystem.
Integrate the option to use an OverlayFS as the root filesystem
into the 90dmsquash-live module for testing purposes.

Tested primarily with transient, in-RAM overlay boots on vfat-
formatted Live USB device, with persistent overlay directories
on ext4-formatted Live USB devices, and with embedded, persistent
overlay directories on vfat-formatted devices. (Persistent overlay
directories on a vfat-formatted device must be in an embedded
filesystem that supports the creation of trusted.* extended attributes,
and must provide valid d_type in readdir responses.)  The
rd.live.overlay.readonly option to allow a persistent overlayfs to be
mounted read only through a higher level transient overlay directory
has been implemented through the multiple lower layers feature of
OverlayFS.

This implementation currently requires the enforcing=0 kernel command
line option.  Further SELinux tuning may change this limitation.

To permit easier comparison of Device-mapper vs. OverlayFS performance,
the default transient dm overlay size has been adjusted up to 32 GiB.
A transient dm overlay is a sparse file in memory, so this does not
consume more RAM for legacy applications.  It does permit a user to
use all of the available root filesystem storage, and fails gently
when it is consumed, as the available free root filesystem storage on
a typical LiveOS build is under a few GiB.  Thus, when booted on other-
than-small RAM systems, the transient dm overlay should not overflow.

OverlayFS offers the potential to use all of the available free RAM,
or all of the available free disc storage on non-vfat-devices,
in its overlay, even beyond the root filesystem available space,
because the OverlayFS root filesystem is a union of directories on
two different partitions.

+++ modules.d/90dmsquash-live/dmsquash-live-root.sh
@@ -28,9 +28,10 @@ getargbool 0 rd.live.overlay.readonly -d -y readonly_overlay && readonly_overlay
 overlay=$(getarg rd.live.overlay -d overlay)
 getargbool 0 rd.writable.fsimg -d -y writable_fsimg && writable_fsimg="yes"
 overlay_size=$(getarg rd.live.overlay.size=)
-[ -z "$overlay_size" ] && overlay_size=512
+[ -z "$overlay_size" ] && overlay_size=32768
 
 getargbool 0 rd.live.overlay.thin && thin_snapshot="yes"
+getargbool 0 rd.live.overlay.overlayfs && overlayfs="yes"
 
 # CD/DVD media check
 [ -b $livedev ] && fs=$(blkid -s TYPE -o value $livedev)
@@ -110,31 +111,54 @@ do_live_overlay() {
     setup=""
     if [ -n "$devspec" -a -n "$pathspec" -a -n "$overlay" ]; then
         mkdir -m 0755 /run/initramfs/overlayfs
+        opt=''
+        [ -n "$readonly_overlay" ] && opt=-r
         mount -n -t auto $devspec /run/initramfs/overlayfs || :
         if [ -f /run/initramfs/overlayfs$pathspec -a -w /run/initramfs/overlayfs$pathspec ]; then
-            losetup $OVERLAY_LOOPDEV /run/initramfs/overlayfs$pathspec
-            if [ -n "$reset_overlay" ]; then
-                dd if=/dev/zero of=$OVERLAY_LOOPDEV bs=64k count=1 conv=fsync 2>/dev/null
+            losetup $opt $OVERLAY_LOOPDEV /run/initramfs/overlayfs$pathspec
+            umount -l /run/initramfs/overlayfs || :
+            oltype=$(det_img_fs $OVERLAY_LOOPDEV)
+            if [ $oltype = DM_snapshot_cow ]; then
+                if [ -n "$reset_overlay" ]; then
+                    dd if=/dev/zero of=$OVERLAY_LOOPDEV bs=64k count=1 conv=fsync 2>/dev/null
+                fi
+            else
+                mount -n -t $oltype $opt $OVERLAY_LOOPDEV /run/initramfs/overlayfs
+                ln -s /run/initramfs/overlayfs/overlayfs /run/overlayfs$opt
+                ln -s /run/initramfs/overlayfs/ovlwork /run/ovlwork$opt
             fi
-            setup="yes"
+        elif [ -d /run/initramfs/overlayfs$pathspec ]; then
+            ln -s /run/initramfs/overlayfs$pathspec /run/overlayfs$opt
+            ln -s /run/initramfs/overlayfs$pathspec/../ovlwork /run/ovlwork$opt
         fi
-        umount -l /run/initramfs/overlayfs || :
+        setup="yes"
+    fi
+    if [ -n "$overlayfs" ]; then
+        modprobe overlay
+        if [ $? != 0 ]; then
+            echo "OverlayFS is not available; using temporary Device-mapper overlay." > /dev/kmsg
+            unset -v overlayfs setup
+        fi
     fi
 
     if [ -z "$setup" -o -n "$readonly_overlay" ]; then
         if [ -n "$setup" ]; then
             warn "Using temporary overlay."
         elif [ -n "$devspec" -a -n "$pathspec" ]; then
-            warn "Unable to find persistent overlay; using temporary"
+            echo "Unable to find persistent overlay; using temporary." > /dev/kmsg
             sleep 5
         fi
-
-        dd if=/dev/null of=/overlay bs=1024 count=1 seek=$((overlay_size*1024)) 2> /dev/null
-        if [ -n "$setup" -a -n "$readonly_overlay" ]; then
-            RO_OVERLAY_LOOPDEV=$( losetup -f )
-            losetup $RO_OVERLAY_LOOPDEV /overlay
+        if [ -n "$overlayfs" ]; then
+            mkdir -m 0755 /run/overlayfs
+            mkdir -m 0755 /run/ovlwork
         else
-            losetup $OVERLAY_LOOPDEV /overlay
+            dd if=/dev/null of=/overlay bs=1024 count=1 seek=$((overlay_size*1024)) 2> /dev/null
+            if [ -n "$setup" -a -n "$readonly_overlay" ]; then
+                RO_OVERLAY_LOOPDEV=$( losetup -f )
+                losetup $RO_OVERLAY_LOOPDEV /overlay
+            else
+                losetup $OVERLAY_LOOPDEV /overlay
+            fi
         fi
     fi
 
@@ -151,7 +175,7 @@ do_live_overlay() {
 
     if [ -n "$thin_snapshot" ]; then
         modprobe dm_thin_pool
-        mkdir /run/initramfs/thin-overlay
+        mkdir -m 0755 /run/initramfs/thin-overlay
 
         # In block units (512b)
         thin_data_sz=$(( $overlay_size * 1024 * 1024 / 512 ))
@@ -170,7 +194,7 @@ do_live_overlay() {
 
         # Create a snapshot of the base image
         echo 0 $sz thin /dev/mapper/live-overlay-pool 0 $base | dmsetup create live-rw
-    else
+    elif [ -z "$overlayfs" ]; then
         echo 0 $sz snapshot $base $over PO 8 | dmsetup create live-rw
     fi
 
@@ -204,10 +228,10 @@ fi
 if [ -e /run/initramfs/live/${live_dir}/${squash_image} ]; then
     SQUASHED="/run/initramfs/live/${live_dir}/${squash_image}"
     if [ -n "$live_ram" ]; then
-        echo "Copying live image to RAM..."
-        echo "(this may take a few minutes)"
+        echo 'Copying live image to RAM...' > /dev/kmsg
+        echo ' (this may take a minute)' > /dev/kmsg
         dd if=$SQUASHED of=/run/initramfs/squashed.img bs=512 2> /dev/null
-        echo "Done copying live image to RAM."
+        echo 'Done copying live image to RAM.' > /dev/kmsg
         SQUASHED="/run/initramfs/squashed.img"
     fi
 
@@ -229,10 +253,10 @@ else
         FSIMG="/run/initramfs/live/${live_dir}/rootfs.img"
     fi
     if [ -n "$live_ram" ]; then
-        echo 'Copying live image to RAM...'
-        echo '(this may take a few minutes)'
+        echo 'Copying live image to RAM...' > /dev/kmsg
+        echo ' (this may take a minute or so)' > /dev/kmsg
         dd if=$FSIMG of=/run/initramfs/rootfs.img bs=512 2> /dev/null
-        echo 'Done copying live image to RAM.'
+        echo 'Done copying live image to RAM.' > /dev/kmsg
         FSIMG='/run/initramfs/rootfs.img'
     fi
 fi
@@ -242,8 +266,8 @@ if [ -n "$FSIMG" ] ; then
 
     if [ -n "$writable_fsimg" ] ; then
         # mount the provided filesystem read/write
-        echo "Unpacking live filesystem (may take some time)"
-        mkdir /run/initramfs/fsimg/
+        echo "Unpacking live filesystem (may take some time)" > /dev/kmsg
+        mkdir -m 0755 /run/initramfs/fsimg/
         if [ -n "$SQUASHED" ]; then
             cp -v $FSIMG /run/initramfs/fsimg/rootfs.img
         else
@@ -251,19 +275,19 @@ if [ -n "$FSIMG" ] ; then
         fi
         FSIMG=/run/initramfs/fsimg/rootfs.img
     fi
-    if [ -n "$writable_fsimg" ] || [ -z "$SQUASHED" -a -n "$live_ram" ] ||
+    if [ -n "$writable_fsimg" ] || [ -z "$SQUASHED" -a -n "$live_ram" -a -z "$overlayfs" ] ||
        [ "$overlay" = none -o "$overlay" = None -o "$overlay" = NONE ]; then
         losetup $BASE_LOOPDEV $FSIMG
         sz=$(blockdev --getsz $BASE_LOOPDEV)
         echo 0 $sz linear $BASE_LOOPDEV 0 | dmsetup create live-rw
     else
-        # mount the filesystem read-only and add a dm snapshot for writes
+        # Mount the filesystem read-only and add a dm snapshot or OverlayFS for writes.
         losetup -r $BASE_LOOPDEV $FSIMG
         do_live_from_base_loop
     fi
 fi
 
-[ -e "$SQUASHED" ] && umount -l /run/initramfs/squashfs
+[ -e "$SQUASHED" ] && [ -z "$overlayfs" ] && umount -l /run/initramfs/squashfs
 
 if [ -b "$OSMIN_LOOPDEV" ]; then
     # set up the devicemapper snapshot device, which will merge
@@ -276,10 +300,26 @@ if [ -n "$ROOTFLAGS" ]; then
     ROOTFLAGS="-o $ROOTFLAGS"
 fi
 
-ln -s /dev/mapper/live-rw /dev/root
-
-if [ -z "$DRACUT_SYSTEMD" ]; then
-    printf 'mount %s /dev/mapper/live-rw %s\n' "$ROOTFLAGS" "$NEWROOT" > $hookdir/mount/01-$$-live.sh
+if [ -n "$overlayfs" ]; then
+    mkdir -m 0755 /run/rootfsbase
+    if [ -n "$reset_overlay" ] && [ -L /run/overlayfs ]; then
+        rm -r -- $(readlink /run/overlayfs)
+        mkdir -m 0755 $(readlink /run/overlayfs)
+    fi
+    if [ -n "$readonly_overlay" ]; then
+        mkdir -m 0755 /run/rootfsbase-r
+        mount -r $FSIMG /run/rootfsbase-r
+        mount -t overlay LiveOS_rootfs-r -oro,lowerdir=/run/overlayfs-r:/run/rootfsbase-r /run/rootfsbase
+    else
+        mount -r $FSIMG /run/rootfsbase
+    fi
+    # Maintain a link at /dev/mapper/live-rw in case legacy code expects it.
+    ln -s /dev/mapper/live-base /dev/mapper/live-rw
+else
+    ln -s /dev/mapper/live-rw /dev/root
+    if [ -z "$DRACUT_SYSTEMD" ]; then
+        printf 'mount %s /dev/mapper/live-rw %s\n' "$ROOTFLAGS" "$NEWROOT" > $hookdir/mount/01-$$-live.sh
+    fi
 fi
 
 need_shutdown
